+++ spec/chewy/type/adapter/active_record_spec.rb
@@ -16,6 +16,14 @@ describe Chewy::Type::Adapter::ActiveRecord, :active_record do
     end
   end
 
+  describe '#default_scope' do
+    specify { expect(described_class.new(City).default_scope).to eq(City.where(nil)) }
+    specify { expect(described_class.new(City.order(:id)).default_scope).to eq(City.where(nil)) }
+    specify { expect(described_class.new(City.limit(10)).default_scope).to eq(City.where(nil)) }
+    specify { expect(described_class.new(City.offset(10)).default_scope).to eq(City.where(nil)) }
+    specify { expect(described_class.new(City.where(rating: 10)).default_scope).to eq(City.where(rating: 10)) }
+  end
+
   describe '#type_name' do
     specify { expect(described_class.new(City).type_name).to eq('city') }
     specify { expect(described_class.new(City.order(:id)).type_name).to eq('city') }
@@ -161,21 +169,29 @@ describe Chewy::Type::Adapter::ActiveRecord, :active_record do
 
       specify { expect(import).to eq([{index: cities.first(3)}]) }
 
-      specify { expect(import(City.order(:id))).to eq([{index: cities.first(3)}]) }
-      specify { expect(import(City.order(:id), batch_size: 2))
-        .to eq([{index: cities.first(2)}, {index: [cities[2]]}]) }
+      specify { expect(import(City.where('rating < 2')))
+        .to eq([{index: cities.first(3)}, {delete: [cities.last.id]}]) }
+      specify { expect(import(City.where('rating < 2'), batch_size: 2))
+        .to eq([{index: cities.first(2)}, {index: [cities[2]]}, {delete: [cities.last.id]}]) }
+      specify { expect(import(City.where('rating < 1')))
+        .to eq([{index: cities.first(3)}]) }
+      specify { expect(import(City.where('rating > 1'))).to eq([]) }
 
+      specify { expect(import(cities.first(2)))
+        .to eq([{index: cities.first(2)}]) }
       specify { expect(import(cities))
-        .to eq([{index: cities.first(3), delete: [cities.last]}]) }
+        .to eq([{index: cities.first(3), delete: cities.last(1)}]) }
       specify { expect(import(cities, batch_size: 2))
-        .to eq([{index: cities.first(2)}, {index: [cities[2]], delete: [cities.last]}]) }
+        .to eq([{index: cities.first(2)}, {index: [cities[2]], delete: cities.last(1)}]) }
       specify { expect(import(cities, deleted))
-        .to eq([{index: cities.first(3), delete: [cities.last] + deleted}]) }
+        .to eq([{index: cities.first(3), delete: cities.last(1) + deleted}]) }
       specify { expect(import(cities, deleted, batch_size: 3)).to eq([
         {index: cities.first(3)},
-        {delete: [cities.last] + deleted.first(2)},
+        {delete: cities.last(1) + deleted.first(2)},
         {delete: deleted.last(1)}]) }
 
+      specify { expect(import(cities.first(2).map(&:id)))
+        .to eq([{index: cities.first(2)}]) }
       specify { expect(import(cities.map(&:id)))
         .to eq([{index: cities.first(3)}, {delete: [cities.last.id]}]) }
       specify { expect(import(cities.map(&:id), batch_size: 2))